MOD: AdoptReviewModel, interfaces 파일로 이동
